refactor(impact-analysis): Moved service to separate package.
