+++ src/main/java/com/synopsys/integration/detect/tool/impactanalysis/VulnerabilityImpactAnalysisTool.java
@@ -6,6 +6,8 @@ import java.nio.file.Path;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.synopsys.integration.detect.tool.impactanalysis.service.ImpactAnalysisService;
+import com.synopsys.integration.detect.tool.impactanalysis.service.ImpactAnalysisUploadResult;
 import com.synopsys.integration.detect.workflow.codelocation.CodeLocationNameManager;
 import com.synopsys.integration.detect.workflow.file.DirectoryManager;
 import com.synopsys.integration.exception.IntegrationException;
