+++ src/main/java/com/synopsys/integration/detect/tool/impactanalysis/service/ImpactAnalysisService.java
@@ -1,4 +1,4 @@
-package com.synopsys.integration.detect.tool.impactanalysis;
+package com.synopsys.integration.detect.tool.impactanalysis.service;
 
 import java.io.File;
 import java.io.IOException;
