+++ src/main/java/com/synopsys/integration/detect/tool/impactanalysis/service/ImpactAnalysisSuccessResult.java
@@ -1,4 +1,4 @@
-package com.synopsys.integration.detect.tool.impactanalysis;
+package com.synopsys.integration.detect.tool.impactanalysis.service;
 
 import com.google.gson.JsonObject;
 import com.google.gson.annotations.SerializedName;
